@@ -1,32 +1,35 @@
 ﻿using SiraUtil.Affinity;
 using System.Collections.Generic;
 using UnityEngine;
-using Zenject;
 
-namespace OverlySensitiveSpectrograms.AffinityPatches
+namespace OverlySensitiveSpectrograms.AffinityPatches;
+
+internal class BasicSpectrogramDataPatch : IAffinity
 {
-    internal class BasicSpectrogramDataPatch : IAffinity
-    {
-        [Inject] private Config _config;
+    private readonly Config _config;
 
-        [AffinityPatch(typeof(BasicSpectrogramData), "ProcessSamples"), AffinityPrefix]
-        private bool ProcessSamples(float[] sourceSamples, List<float> processedSamples)
+    public BasicSpectrogramDataPatch(Config config)
+    {
+        _config = config;
+    }
+
+    [AffinityPatch(typeof(BasicSpectrogramData), nameof(BasicSpectrogramData.ProcessSamples)), AffinityPrefix]
+    private bool ProcessSamplesPrefix(float[] sourceSamples, List<float> processedSamples)
+    {
+        var deltaTime = Time.deltaTime;
+        for (var i = 0; i < sourceSamples.Length; i++)
         {
-            var deltaTime = Time.deltaTime;
-            for (var i = 0; i < sourceSamples.Length; i++)
+            var sample = Mathf.Log(sourceSamples[i] + 1f) * (i + 1) * _config.SampleBoost;
+            if (processedSamples[i] < sample)
             {
-                var sample = Mathf.Log(sourceSamples[i] + 1f) * (i + 1) * _config.SampleBoost;
-                if (processedSamples[i] < sample)
-                {
-                    if (_config.UseInstantChangeThreshold && sample - processedSamples[i] > _config.InstantChangeThreshold)
-                        processedSamples[i] = sample;
-                    else
-                        processedSamples[i] = Mathf.Lerp(processedSamples[i], sample, deltaTime * _config.SampleLerp0);
-                }
+                if (_config.UseInstantChangeThreshold && sample - processedSamples[i] > _config.InstantChangeThreshold)
+                    processedSamples[i] = sample;
                 else
-                    processedSamples[i] = Mathf.Lerp(processedSamples[i], sample, deltaTime * _config.SampleLerp1);
+                    processedSamples[i] = Mathf.Lerp(processedSamples[i], sample, deltaTime * _config.SampleLerp0);
             }
-            return false;
+            else
+                processedSamples[i] = Mathf.Lerp(processedSamples[i], sample, deltaTime * _config.SampleLerp1);
         }
+        return false;
     }
 }
