@@ -6,24 +6,23 @@ using SiraUtil.Zenject;
 using IPAConfig = IPA.Config.Config;
 using IPALogger = IPA.Logging.Logger;
 
-namespace OverlySensitiveSpectrograms
+namespace OverlySensitiveSpectrograms;
+
+[Plugin(RuntimeOptions.DynamicInit), NoEnableDisable, Slog]
+public class Plugin
 {
-    [Plugin(RuntimeOptions.DynamicInit), NoEnableDisable, Slog]
-    public class Plugin
+    [Init]
+    public Plugin(IPALogger logger, IPAConfig conf, Zenjector zenjector)
     {
-        [Init]
-        public Plugin(IPALogger logger, IPAConfig conf, Zenjector zenjector)
+        zenjector.UseLogger(logger);
+
+        var config = conf.Generated<Config>();
+        zenjector.Install(Location.App, Container =>
         {
-            zenjector.UseLogger(logger);
+            Container.BindInstance(config).AsSingle();
+        });
 
-            var config = conf.Generated<Config>();
-            zenjector.Install(Location.App, container =>
-            {
-                container.BindInstance(config).AsSingle();
-            });
-
-            zenjector.Install<OSSMenuInstaller>(Location.Menu);
-            zenjector.Install<OSSGameCoreInstaller>(Location.GameCore);
-        }
+        zenjector.Install<OSSMenuInstaller>(Location.Menu);
+        zenjector.Install<OSSGameCoreInstaller>(Location.GameCore);
     }
 }
